Made ARAM-blocks cache have 8x4kb MEM1 cached blocks, fixes Dynarec performance on GC. Fixed ROM-Cache.c function header.

+++ trunk/main/ROM-Cache.c
@@ -159,7 +159,7 @@ void ARAM_ReadFromBlock(char *block,int startOffset, int bytes, char *dest)
 	free(buffer);
 }
 
-void ROMCache_read(u32* ram_dest, u32 rom_offset, u32 length){
+void ROMCache_read(u8* ram_dest, u32 rom_offset, u32 length){
 
 	if(ROM_too_big){ // The whole ROM isn't in ARAM, we might have to move blocks in/out
 		u32 length2 = length;
